Move #comment from CtStatementImpl to CtElementImpl Remove #comment from CtClassImpl, CtUnaryOperatorImpl, CtInvocationImpl and CtConstructorCallImpl Based on comments and suggestions at https://github.com/INRIA/spoon/pull/3494 Fixes #2013
