+++ src/main/java/spoon/support/reflect/code/CtUnaryOperatorImpl.java
@@ -8,7 +8,6 @@
 package spoon.support.reflect.code;
 
 import spoon.reflect.annotations.MetamodelPropertyField;
-import spoon.reflect.code.CtComment;
 import spoon.reflect.code.CtExpression;
 import spoon.reflect.code.CtStatement;
 import spoon.reflect.code.CtStatementList;
@@ -103,20 +102,5 @@ public class CtUnaryOperatorImpl<T> extends CtExpressionImpl<T> implements CtUna
 	@Override
 	public CtUnaryOperator<T> clone() {
 		return (CtUnaryOperator<T>) super.clone();
-	}
-
-	@Override
-	public void comment() {
-		if (!isParentInitialized()) {
-			// already not in a tree, commenting wouldn't make a difference
-			return;
-		}
-		// comment is implemented as replace by a comment
-		final String stmt = toString();
-		if (stmt.contains(CtComment.LINE_SEPARATOR)) {
-			this.replace(getFactory().Code().createComment(stmt, CtComment.CommentType.BLOCK)); // Multi line comment
-		} else {
-			this.replace(getFactory().Code().createInlineComment(stmt + ';')); // Single line comment
-		}
 	}
 }
